// src/app/Interfaces/v-gc-faltas-dto.ts
export interface VGCFaltasDTO {
    iCveRegistroEvento?: number;
    iCveUnica: number;
    iCveRegistroPersona?: number;
    cCURP?: string;
    cPrimerApellido?: string;
    cSegundoApellido?: string;
    cNombre?: string;
    iCveMotivoFalta?: number;
    cDescMotivoFalta?: string;
    cHospital?: string;
    cObservacion?: string;
    iCveEstatusEvento?: number;
    dFechaEvento?: Date;
}